updated readme, moved a few files around
